refactor(gdb): re-structure lvglgdb module
